requirements.txt
python-telegram-bot==5.0.0
pytz==2016.6.1
rethinkdb==2.3.0.post6
khayyam==3.0.15
